refactor(contexts): split Lightbox, SidebarLayout, Tabs, Settings providers for Fast Refresh
A file that exports a component plus a hook/context bails out of React
Fast Refresh. Move each context object and consumer hook into a
component-free `XContext.ts` module and keep only the provider component
in `XProvider.tsx`. Mirrors the SyncConfig split so every provider is
Fast-Refresh-eligible.
