+++ src/components/App.tsx
@@ -1,6 +1,6 @@
-import { SidebarLayoutProvider } from "@/contexts/SidebarLayoutContext";
+import { SidebarLayoutProvider } from "@/contexts/SidebarLayoutProvider";
 import { SyncConfigProvider } from "@/contexts/SyncConfigProvider";
-import { TabsProvider } from "@/contexts/TabsContext";
+import { TabsProvider } from "@/contexts/TabsProvider";
 import { useCodeThemeStyle } from "@/hooks/useCodeThemeStyle";
 import { useLocale } from "@/hooks/useLocale";
 import { useSettings } from "@/hooks/useSettings";
