+++ src/contexts/TabsProvider.tsx
@@ -1,28 +1,12 @@
-import { createContext, type ReactNode, useContext, useMemo } from "react";
+import { type ReactNode, useMemo } from "react";
 import { useSettings } from "@/hooks/useSettings";
-import { type TocEntry, useTableOfContents } from "@/hooks/useTableOfContents";
+import { useTableOfContents } from "@/hooks/useTableOfContents";
 import { useTabs } from "@/hooks/useTabs";
-import { useWorkspaceNotice, type WorkspaceNotice } from "@/hooks/useWorkspaceNotice";
-import { type Backlink, filterBacklinks } from "@/lib/backlinks";
+import { useWorkspaceNotice } from "@/hooks/useWorkspaceNotice";
+import { filterBacklinks } from "@/lib/backlinks";
 import { displayContentFor, tocContentFor } from "@/lib/displayContent";
 import { EDITOR_MODE } from "@/lib/settings";
-
-type TabsApi = ReturnType<typeof useTabs>;
-
-export interface TabsContextValue extends TabsApi {
-  // Derived from the active file + edit mode. View mode renders saved content;
-  // edit/split renders the in-memory editContent so preview reflects typing.
-  displayContent: string | null;
-  tocEntries: TocEntry[];
-  backlinks: Backlink[];
-  // Notice shown for a workspace event (#262): a refusal, or a persistent
-  // warning when a folder is opened inside a parent git repo. A translation
-  // key + values so the banner re-localizes live (see WorkspaceNoticeBanner).
-  workspaceNotice: WorkspaceNotice | null;
-  dismissWorkspaceNotice: () => void;
-}
-
-export const TabsContext = createContext<TabsContextValue | null>(null);
+import { TabsContext, type TabsContextValue } from "./TabsContext";
 
 export function TabsProvider({ children }: { children: ReactNode }) {
   const { settings, updateSettings } = useSettings();
@@ -77,19 +61,4 @@ export function TabsProvider({ children }: { children: ReactNode }) {
   );
 
   return <TabsContext.Provider value={value}>{children}</TabsContext.Provider>;
-}
-
-export function useTabsContext(): TabsContextValue {
-  const ctx = useContext(TabsContext);
-  if (!ctx) throw new Error("useTabsContext must be used inside <TabsProvider>");
-  return ctx;
-}
-
-// The opened workspace folder's root (or undefined when only loose files are
-// open), read straight from the tabs context. Uses optional chaining rather
-// than useTabsContext so the shared renderers (notebooks, canvas, print, and
-// isolated tests) can read it with no provider and simply skip workspace
-// resolution.
-export function useWorkspaceRoot(): string | undefined {
-  return useContext(TabsContext)?.workspace?.root;
 }
